changed: data type

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -3,16 +3,16 @@
 /**
  * list_len - calculates the num of nodes
  * 
- * @h: single linked list
+ * @h:single linked list
  * 
- * Return: count of linked lists
+ * Return:count of linked lists
  * 
  */
 
 
 size_t list_len(const list_t *h){
 
-    int count = 0;
+    size_t count = 0;
 
     while (h != NULL)
     {
